+++ core/src/components/image/bn-image-imaginary-provider.tsx
@@ -1,4 +1,4 @@
-import { Component, ComponentInterface, Element, h, Host, Prop } from "@stencil/core";
+import { Component, ComponentInterface, Element, h, Host, Method, Prop } from "@stencil/core";
 import { imageProvider } from "../../utils/imageProvider";
 import { Logger } from "../../utils/utils";
 import { ImaginaryTransform } from "./bn-image-imaginary-provider.const";
@@ -14,11 +14,28 @@ export class BnImageImaginaryProvider implements ComponentInterface {
 
   @Prop({ reflect: true }) readonly origin!: string;
 
+  private _transform?: ImaginaryTransform;
+  private _getTransform() {
+    if (this._transform?.config.origin !== this.origin) {
+      this._transform = new ImaginaryTransform({ origin: this.origin });
+    }
+    return this._transform!;
+  }
+
   connectedCallback() {
-    imageProvider.registryTransform(this.hostEle, new ImaginaryTransform({ origin: this.origin }));
+    imageProvider.registryTransform(this.hostEle, this._getTransform());
   }
   disconnectedCallback() {
     imageProvider.unregistryTransform(this.hostEle);
+  }
+
+  @Method()
+  async transform(src: string, params: { [key: string]: unknown }) {
+    return this._getTransform().transform(src, params);
+  }
+  @Method()
+  async transformFromElement(src: string, ele: HTMLElement) {
+    return this._getTransform().transformFromElement(src, ele);
   }
 
   render() {