+++ core/src/utils/imageProvider.ts
@@ -1,5 +1,11 @@
 export abstract class ImageTransform {
-  abstract transform(source_url: string, ele: HTMLElement): /* dest_url */ string | PromiseLike<string>;
+  abstract transform(
+    source_url: string,
+    params: { [key: string]: unknown },
+  ): /* dest_url */ string | PromiseLike<string>;
+  transformFromElement(source_url: string, ele: HTMLElement) {
+    return this.transform(source_url, ele.dataset);
+  }
 }
 
 class NoImageTransform extends ImageTransform {
@@ -13,8 +19,8 @@ export const noImageTransform = new NoImageTransform();
  * 未来可能会支持让 bn-image 选择 provider（通过id绑定）：通过 EventBindingHelper 来让 bn-image 与 Provider 进行联动
  */
 class ImageProvider extends ImageTransform {
-  transform(source_url: string, ele: HTMLElement) {
-    return this._dtf.transform(source_url, ele);
+  transform(source_url: string, params: { [key: string]: unknown }) {
+    return this._dtf.transform(source_url, params);
   }
   private _rtsMap = new Map<HTMLElement, ImageTransform>();
   private _dtf: ImageTransform = noImageTransform;